Extract DbContext.Get{FullTable,Column}Name extension methods and move them to LeanCode.Dapper
Also move related tests.
