+++ LeanCode.CoreLibrary.sln
@@ -179,6 +179,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.AzureIdentity", "s
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.AzureIdentity.Tests", "test\Infrastructure\LeanCode.AzureIdentity.Tests\LeanCode.AzureIdentity.Tests.csproj", "{BA83956A-6867-4E63-9417-C227F477981B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.Dapper.Tests", "test\Infrastructure\LeanCode.Dapper.Tests\LeanCode.Dapper.Tests.csproj", "{8DF10985-C8B8-4F30-ACD4-3A1110D53550}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1080,6 +1082,18 @@ Global
 		{BA83956A-6867-4E63-9417-C227F477981B}.Release|x64.Build.0 = Release|Any CPU
 		{BA83956A-6867-4E63-9417-C227F477981B}.Release|x86.ActiveCfg = Release|Any CPU
 		{BA83956A-6867-4E63-9417-C227F477981B}.Release|x86.Build.0 = Release|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Debug|x64.Build.0 = Debug|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Debug|x86.Build.0 = Debug|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Release|x64.ActiveCfg = Release|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Release|x64.Build.0 = Release|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Release|x86.ActiveCfg = Release|Any CPU
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{7AFA22EA-B093-4457-ABA2-CFF96DF41350} = {7324BABF-C1F7-430D-BF76-CF9BFA559DC6}
@@ -1166,5 +1180,6 @@ Global
 		{0E5786B1-DCA4-441D-8165-C884E62D7AB9} = {9BCF1002-8885-4598-8E7F-BD491BB5A6B6}
 		{0E3A75B7-8064-4288-836E-CD695177633D} = {1358A1B6-E8BB-4607-8B61-FDAE49F0EAA4}
 		{BA83956A-6867-4E63-9417-C227F477981B} = {D9CFC5C4-8B87-4EDF-85D1-1471EA1B6D27}
+		{8DF10985-C8B8-4F30-ACD4-3A1110D53550} = {D9CFC5C4-8B87-4EDF-85D1-1471EA1B6D27}
 	EndGlobalSection
 EndGlobal
